:lipstick: Fix file
PostUpload에서 사용 되는 style 파일 변경 Fixes: #41
